cart: fix promos not affecting shipping costs (#80)
* cart: Add Sum function to AppliedDiscounts

* cart: Replace DiscountAmount with AppliedDiscounts for ShippingItem

* cart: ShippingItem implements WithDiscount

* cart: Update Changelog

* cart: Assert that ShippingItem implements WithDiscount

* cart: Helper functions mergeDiscountsOnItemLevel and hasDiscountsApplied

Also removed the call on `discountable.MergeDiscounts()` inside `mapDiscounts()` because the latter is a helper function for `WithDiscount.MergeDiscount()`. `Item.MergeDiscount()` and `ShippingItem.MergeDiscount()` could previously not use `mapDiscounts` because they could not pass a valid `WithDiscount` (child) object. Because of that these two functions also did not return what was expected from them (they returned only sorted discounts but not aggregated discount). This is also fixed with this commit.

* cart: Update comments to adhere to golint rules

* cart: Revert changes to mapDiscount()