checkout: optional cart validation before place order
When using the place order state machine, one of the first steps during the process is already validating the cart.
Afterwards the cart is closed and stored inside the context store.
When then later placing the order the OrderService performs the cart validation again.

At this time the cart is already closed and if a validator tries to fetch the cart from the session, an empty cart returns which can lead to invalid validation results.