product: Add concept of LoyaltyEarning
Currently the product domain model only supports paying with loyalty programs, it would be nice to also support a potential earning of different loyalty points/currencies.
This adds LoyaltyEarning properties to the product model which then can later be used to display on the product detail page or in the list view.
Furthermore there is a function added to get the loyalty rate to later change the earning based on the product price after discounts etc.