checkout: In case of error during payment always redirect back to checkout/review page in stead of rendering the template under a different route
Currently in case of a error during payment/placing the order we just render the checkout / review template again on the same route.

If the user then reloads the page he is still on the placeorder route and will trigger the error again or cause other potential side effects.
Its therefore better to send him back on the checkout route and render the error in the matching action. So that the customer then can start a new try.